default for time & moved language check to helper.py
