+++ mysite/polls/helper.py
@@ -1,7 +1,16 @@
 import yaml
+from django.utils.translation import gettext_lazy as _
+from django.core.exceptions import ValidationError
 
 BLACKLISTED_WORDS = 'polls/blacklist.yaml'
 FORTUNE = 'polls/import_data.yaml'
+
+
+def language_filter(text):
+    value = language_check(text)
+    if value[0]:
+        raise ValidationError(_(
+                'Coarse words like ' + value[1] + ' are not allowed.'))
 
 
 def load_yaml_blacklist():
